added UDF test in sanity.py rather then seperately
